Edited .travis.yml added convergence_test.py(visual checks)
