+++ .travis.yml
@@ -2,11 +2,13 @@ sudo: required
 dist: trusty
 language: python
 
+
 python:
     - "3.5"
 
 # command to install dependencies
 before_install:
+    - sudo apt-get install cmake
     - ls -al
     - pwd
     - cd ..
